+++ src/render.ts
@@ -1,11 +1,12 @@
 import Component from './component/component';
+import { renderedHook } from './utils/hooks';
 import { edelweissPolicy } from './utils/trusted_types';
 import { eventListenersMap } from './template/template';
 import { arrayFrom, tupleOf } from '@fluss/core';
 import { loadCSS, unloadCSS } from './utils/styles';
+import { querySelector, cloneNode } from '@fluss/web';
 import { stylePaths, stylePathsToRemove } from './css';
 import { diff, normalizeHTML, attachEvents } from './utils/dom';
-import { querySelector, replaceNode, cloneNode } from '@fluss/web';
 
 /** Render templates on the page. */
 export function render(
@@ -40,10 +41,14 @@ export function render(
           stylePathsToRemove.clear();
           return element;
         })
-        .then((element) => {
-          toElement.hasChildNodes()
-            ? diff(toElement, element)
-            : replaceNode(toElement, element);
+        .then((element) => diff(toElement, element))
+        .then(() => {
+          /**
+           * This trick is used in order to render "initial" state
+           * of page and then apply rendered hook. This is the difference
+           * between mounted and rendered hook.
+           */
+          setTimeout(() => renderedHook(toElement), 0);
         });
     })
     .extract();