+++ src/utils/dom.ts
@@ -1,6 +1,8 @@
 import Component from '../component/component';
+import { isTextNode } from './predicates';
 import { eventListenersMap } from '../template/template';
 import { dataEventIdJSRegExp } from './regexps';
+import { mountedHook, removedHook, updatedHook } from './hooks';
 import { maybeOf, promiseOf, arrayFrom, isNothing } from '@fluss/core';
 import {
   removeNode,
@@ -8,6 +10,7 @@ import {
   replaceNode,
   hasAttribute,
   setAttribute,
+  getAttribute,
   removeAttribute,
   addEventListener,
 } from '@fluss/web';
@@ -15,7 +18,7 @@ import {
 export function diff(oldNode: Element, newNode: Element) {
   if (oldNode.nodeType === newNode.nodeType) {
     if (oldNode.tagName === newNode.tagName) {
-      diffAttributes(oldNode, newNode);
+      diffAttributes(oldNode, newNode) && updatedHook(oldNode);
 
       if (newNode.hasChildNodes()) {
         const oldChildNodes = arrayFrom(oldNode.childNodes);
@@ -31,52 +34,74 @@ export function diff(oldNode: Element, newNode: Element) {
 
           if (!isNothing(nNode)) {
             isNothing(oNode)
-              ? appendNodes(oldNode, nNode)
-              : nNode.nodeType === Node.TEXT_NODE &&
-                oNode.nodeType === Node.TEXT_NODE
+              ? (appendNodes(oldNode, nNode), mountedHook(nNode))
+              : isTextNode(nNode) && isTextNode(oNode)
               ? // Update text node only if there is difference
                 oNode.textContent !== nNode.textContent
-                ? (oNode.textContent = nNode.textContent)
+                ? ((oNode.textContent = nNode.textContent),
+                  // If Text node is changed, assumes that parent Node is updated
+                  updatedHook(oldNode))
                 : null
               : diff(oNode as Element, nNode as Element);
           } else if (!isNothing(oNode)) {
             removeNode(oNode);
+            removedHook(oNode);
           } else {
             // Do nothing - old and new node is null or undefined.
           }
         }
       } else if (oldNode.hasChildNodes()) {
         replaceNode(oldNode, newNode);
+        mountedHook(newNode);
+        removedHook(oldNode);
       } else {
         // Do nothing - both nodes haven't children and difference of attributes is
         // already checked.
       }
     } else {
       replaceNode(oldNode, newNode);
+      mountedHook(newNode);
+      removedHook(oldNode);
     }
   } else {
     replaceNode(oldNode, newNode);
+    mountedHook(newNode);
+    removedHook(oldNode);
   }
 }
 
 /**
  * Determines attribute that is outdated and update or remove it.
  * Attributes may be in inexact order, so we need go twice on them.
+ * Returns value that tells whether attributes in oldNode and newNode
+ * are different or not.
  */
-function diffAttributes(oldNode: Element, newNode: Element) {
+function diffAttributes(oldNode: Element, newNode: Element): boolean {
+  let areAttributesDifferent = false;
+
   if (oldNode.attributes.length !== newNode.attributes.length) {
     // Remove exessive attributes
     arrayFrom(oldNode.attributes).forEach(({ name }) => {
       if (!hasAttribute(newNode, name)) {
         removeAttribute(oldNode, name);
+        areAttributesDifferent = true;
       }
     });
   }
 
   // Add missing attributes and update changed
-  arrayFrom(newNode.attributes).forEach(({ name, value }) =>
-    setAttribute(oldNode, name, value)
-  );
+  arrayFrom(newNode.attributes).forEach(({ name, value }) => {
+    const maybeOldAttribute = getAttribute(oldNode, name);
+    if (
+      maybeOldAttribute.isNothing() ||
+      maybeOldAttribute.extract() !== value
+    ) {
+      setAttribute(oldNode, name, value);
+      areAttributesDifferent = true;
+    }
+  });
+
+  return areAttributesDifferent;
 }
 
 export async function normalizeHTML(