+++ src/utils/hooks.ts
@@ -1,0 +1,37 @@
+import { isElement } from './predicates';
+import { getAttribute } from '@fluss/web';
+import { Hooks, hooksManager } from '../template/template';
+import { arrayFrom, maybeOf, promiseOf, tupleOf } from '@fluss/core';
+
+/** If parent node is mounted, so its children are also mounted. */
+export function mountedHook(node: Node) {
+  applyHook(node, Hooks.Mounted);
+  arrayFrom(node.childNodes).forEach(mountedHook);
+}
+/** If parent node is rendered, so its children are also rendered. */
+export function renderedHook(node: Node) {
+  applyHook(node, Hooks.Rendered);
+  arrayFrom(node.childNodes).forEach(renderedHook);
+}
+
+export function updatedHook(node: Node) {
+  applyHook(node, Hooks.Updated);
+}
+
+/** If parent node is removed, so its children are also removed. */
+export function removedHook(node: Node) {
+  applyHook(node, Hooks.Removed);
+  arrayFrom(node.childNodes).forEach(removedHook);
+}
+
+function applyHook(node: Node, type: Hooks) {
+  if (isElement(node)) {
+    getAttribute(node, `data-${type}-hook-id`)
+      .map((id) => tupleOf(id, maybeOf(hooksManager[type].get(id))))
+      .map(([id, maybeFn]) =>
+        maybeFn.map((fn) =>
+          promiseOf(fn(node)).then(() => hooksManager[type].delete(id))
+        )
+      );
+  }
+}