+++ src/utils/regexps.ts
@@ -1,5 +1,9 @@
-export const eventListenerRegExp = /@([\w-]+)=['"]?$/
-export const dataEventIdJSRegExp = /^eventId[\d]{1,}$/
-export const customElementRegExp = /<([\w-]+)\s*(:[\w]+)?\s*=$/
-export const styleAttributeRegExp = /style=['"]?[\w\s:;()-]*$/
-export const booleanAttributeRegExp = /\?([\w-]+)=['"]?$/
+export const eventListenerRegExp = /@([\w-]+)=['"]?$/;
+
+export const dataEventIdJSRegExp = /^eventId[\d]{1,}$/;
+
+export const customElementRegExp = /<([\w-]+)\s*(:[\w]+)?\s*=$/;
+
+export const booleanAttributeRegExp = /\?([\w-]+)=['"]?$/;
+
+export const hookAttributeRegExp = /:(mounted|rendered|updated|removed)=$/;